+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/k8s_attributes.py
@@ -55,6 +55,13 @@ K8S_CONTAINER_RESTART_COUNT = "k8s.container.restart_count"
 Number of times the container was restarted. This attribute can be used to identify a particular container (running or stopped) within a container spec.
 """
 
+K8S_CONTAINER_STATUS_LAST_TERMINATED_REASON = (
+    "k8s.container.status.last_terminated_reason"
+)
+"""
+Last terminated reason of the Container.
+"""
+
 K8S_CRONJOB_NAME = "k8s.cronjob.name"
 """
 The name of the CronJob.